+++ apps/web/app/future/video/meeting-ended/[uid]/page.tsx
@@ -1,9 +1,7 @@
-import Page from "@pages/video/meeting-ended/[uid]";
+import Page, { getServerSideProps } from "@pages/video/meeting-ended/[uid]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
-
-import { getServerSideProps } from "@lib/video/meeting-ended/[uid]/getServerSideProps";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
